# api-fuzz/openapi.yaml
openapi: 3.0.3
info:
  title: maint-rust API (fuzz surface)
  version: "0.1.0"
  description: >
    Hand-authored spec covering the GET read surface + the sync/reference write
    endpoints for schemathesis property fuzzing. Not exhaustive of request
    bodies — schemathesis's value here is hammering every GET with generated
    query params / path ids and asserting the server never 500s, never returns
    a non-JSON body, and honors its own error envelope. The two owned write
    paths (/sync/push, /vehicle-classes) carry minimal schemas so stateful
    fuzzing can reach them.
servers:
  - url: http://127.0.0.1:8090/api/maint
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /stock/movements:
    get:
      summary: Stock in/out ledger
      parameters:
        - { name: kind, in: query, required: false, schema: { type: string, enum: [tire, oil] } }
        - { name: from, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer } }
        - { name: offset, in: query, required: false, schema: { type: integer } }
      responses: { "200": { description: ok }, "400": { description: bad } }
  /stock/tires/credit:
    post:
      summary: Credit tire stock (perm >= 4, idempotent on ref_id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brand, qty]
              properties:
                brand: { type: string, example: "Michelin" }
                model: { type: string, nullable: true }
                size: { type: string, nullable: true }
                qty: { type: integer, example: 4 }
                ref_id: { type: string, format: uuid, nullable: true }
      responses: { "200": { description: ok }, "403": { description: forbidden }, "400": { description: bad } }
  /stock/oil/credit:
    post:
      summary: Credit oil stock (perm >= 4, idempotent on ref_id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oil_type, liters]
              properties:
                oil_type: { type: string, example: "15W40" }
                liters: { type: number, example: 200 }
                ref_id: { type: string, format: uuid, nullable: true }
      responses: { "200": { description: ok }, "403": { description: forbidden }, "400": { description: bad } }
  /stock/debit:
    post:
      summary: Debit stock (idempotent on ref_id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, ref_id]
              properties:
                kind: { type: string, enum: [tire_debit, oil_debit] }
                brand: { type: string, nullable: true }
                model: { type: string, nullable: true }
                size: { type: string, nullable: true }
                oil_type: { type: string, nullable: true }
                qty: { type: integer, nullable: true }
                liters: { type: number, nullable: true }
                ref_id: { type: string, format: uuid }
                work_order_id: { type: string, format: uuid, nullable: true }
      responses: { "200": { description: ok }, "400": { description: bad } }
...