# src/handlers/stock.rs
//! Stock caches — reads for the garage, credits for the dashboard (level 4).
//!
//! Per §4 the authoritative credit ledger is Falcon's; these caches mirror it.
//! For local/standalone running, credits write directly here; a real deploy
//! reconciles with Falcon server-side.

use actix_web::{get, post, web, HttpResponse};
use serde::Deserialize;
use serde_json::{json, Value};

use crate::auth::middleware::BearerToken;
use crate::auth::AuthClaims;
use crate::db::PgPool;
use crate::error::{ApiError, ApiResult};
use crate::handlers::AppState;
use crate::services::stock_ledger;

/// Mirror Falcon's authoritative tire ledger into tire_stock_cache, then return it (§4).
#[get("/stock/tires")]
async fn tire_stock(state: web::Data<AppState>, token: BearerToken) -> ApiResult<HttpResponse> {
    // Replay any unmirrored debits FIRST — the refresh below overwrites our
    // local decrements with Falcon's counts, so Falcon must see the debits
    // before its numbers become ours.
    stock_ledger::mirror_unmirrored(&state, &token.0).await;
    if let Ok(v) = state.falcon.get_json("/api/maint-stock/tires", &token.0).await {
        if let Some(arr) = v.as_array() {
            for row in arr {
                let brand = row.get("brand").and_then(|x| x.as_str()).unwrap_or("");
                let model = row.get("model").and_then(|x| x.as_str()).unwrap_or("");
                let size = row.get("size").and_then(|x| x.as_str()).unwrap_or("");
                let qty = row.get("on_hand_qty").and_then(|x| x.as_i64()).unwrap_or(0) as i32;
                let _ = sqlx::query(
                    "INSERT INTO tire_stock_cache (brand, model, size, on_hand_qty) VALUES ($1,$2,$3,$4) \
                     ON CONFLICT (brand, model, size) DO UPDATE SET on_hand_qty = EXCLUDED.on_hand_qty, fetched_at = now()",
                ).bind(brand).bind(model).bind(size).bind(qty).execute(&state.pool).await;
            }
        }
    }
    let rows: Vec<(Value,)> = sqlx::query_as(
        "SELECT to_jsonb(s) FROM tire_stock_cache s ORDER BY brand, model, size",
    ).fetch_all(&state.pool).await?;
    Ok(HttpResponse::Ok().json(rows.into_iter().map(|(v,)| v).collect::<Vec<_>>()))
}

#[get("/stock/oil")]
async fn oil_stock(state: web::Data<AppState>, token: BearerToken) -> ApiResult<HttpResponse> {
    // Same ordering rule as /stock/tires: debits reach Falcon before Falcon's
    // counts replace the local cache.
    stock_ledger::mirror_unmirrored(&state, &token.0).await;
    if let Ok(v) = state.falcon.get_json("/api/maint-stock/oil", &token.0).await {
        if let Some(arr) = v.as_array() {
            for row in arr {
                let oil = row.get("oil_type").and_then(|x| x.as_str()).unwrap_or("");
                let liters = row.get("liters_on_hand").and_then(|x| x.as_f64()).unwrap_or(0.0);
                let _ = sqlx::query(
                    "INSERT INTO oil_stock_cache (oil_type, liters_on_hand) VALUES ($1,$2) \
                     ON CONFLICT (oil_type) DO UPDATE SET liters_on_hand = EXCLUDED.liters_on_hand, fetched_at = now()",
                ).bind(oil).bind(liters).execute(&state.pool).await;
            }
        }
    }
    let rows: Vec<(Value,)> =
        sqlx::query_as("SELECT to_jsonb(s) FROM oil_stock_cache s ORDER BY oil_type")
            .fetch_all(&state.pool).await?;
    Ok(HttpResponse::Ok().json(rows.into_iter().map(|(v,)| v).collect::<Vec<_>>()))
}

/// Debit stock: record the movement locally, decrement the cache, and push the
/// debit up to Falcon (§4). Idempotent upstream on ref_id.
#[derive(Debug, Deserialize)]
struct DebitInput {
    kind: String,       // tire_debit | oil_debit
    brand: Option<String>,
    model: Option<String>,
    size: Option<String>,
    oil_type: Option<String>,
    qty: Option<i32>,
    liters: Option<f64>,
    ref_id: String,     // the mount/oil-change UUID (idempotency key)
    work_order_id: Option<String>,
}

#[post("/stock/debit")]
async fn debit_stock(
    state: web::Data<AppState>,
    token: BearerToken,
    body: web::Json<DebitInput>,
) -> ApiResult<HttpResponse> {
    let d = body.into_inner();
    let ref_uuid = uuid::Uuid::parse_str(&d.ref_id)
        .map_err(|_| ApiError::BadRequest("ref_id must be a uuid".into()))?;

    let kind: &'static str = match d.kind.as_str() {
        "tire_debit" => "tire_debit",
        "oil_debit" => "oil_debit",
        other => {
            return Err(ApiError::BadRequest(format!("unknown debit kind: {other}")));
        }
    };
    let spec = stock_ledger::DebitSpec {
        kind,
        brand: d.brand,
        model: d.model,
        size: d.size,
        oil_type: d.oil_type,
        qty: d.qty,
        liters: d.liters,
        ref_id: ref_uuid,
        work_order_id: d.work_order_id.as_ref().and_then(|s| uuid::Uuid::parse_str(s).ok()),
    };

    // Local movement + cache decrement (idempotent on ref_id).
    let mut tx = state.pool.begin().await?;
    stock_ledger::record_debit(&mut tx, &spec).await?;
    tx.commit().await?;

    // Push up to Falcon (best-effort; idempotent on ref_id upstream).
    stock_ledger::mirror_debits(&state, &token.0, &[ref_uuid]).await;
    Ok(HttpResponse::Ok().json(json!({"ok": true})))
}

#[derive(Debug, Deserialize)]
struct TireCredit {
    brand: String,
    model: Option<String>,
    size: Option<String>,
    qty: i32,
    /// Idempotency key. A retried/double-clicked credit with the same ref_id is
    /// a no-op (returns the current on-hand). Absent = a fresh, non-idempotent
    /// credit (server mints one). The movement row records the same key.
    ref_id: Option<String>,
}

#[post("/stock/tires/credit")]
async fn credit_tires(
    pool: web::Data<PgPool>,
    claims: AuthClaims,
    body: web::Json<TireCredit>,
) -> ApiResult<HttpResponse> {
    if claims.permission < 4 {
        return Err(ApiError::Forbidden("stock credit requires permission >= 4".into()));
    }
    let b = body.into_inner();
    if b.qty <= 0 {
        return Err(ApiError::BadRequest("qty must be > 0".into()));
    }
    let ref_id = parse_opt_uuid(b.ref_id.as_deref())?;

    // Race-free idempotency: insert the movement FIRST with ON CONFLICT DO
    // NOTHING on the unique ref_id (0018). If nothing was inserted, this ref_id
    // already applied — return the current on-hand without bumping. Cache bump
    // + movement in ONE transaction.
    let mut tx = pool.begin().await?;
    let logged: Option<(uuid::Uuid,)> = sqlx::query_as(
        "INSERT INTO stock_movements (kind, brand, model, size, qty, ref_id) \
         VALUES ('tire_credit', $1, $2, $3, $4, COALESCE($5, gen_random_uuid())) \
         ON CONFLICT (ref_id) WHERE ref_id IS NOT NULL DO NOTHING RETURNING id",
    )
    .bind(&b.brand)
    .bind(&b.model)
    .bind(&b.size)
    .bind(b.qty)
    .bind(ref_id)
    .fetch_optional(&mut *tx)
    .await?;
    if logged.is_none() {
        // Idempotency is keyed on ref_id alone — a replay may carry a different
        // brand/model/size than the original (a client that reused a key). Read
        // this request's key with fetch_OPTIONAL and synthesize a zero-on-hand
        // row when it doesn't exist, so a mismatched replay is a benign no-op,
        // never a RowNotFound → 500.
        let row: Option<(Value,)> = sqlx::query_as(
            "SELECT to_jsonb(t) FROM tire_stock_cache t \
             WHERE brand=$1 AND model=COALESCE($2,'') AND size=COALESCE($3,'')",
        )
        .bind(&b.brand)
        .bind(&b.model)
        .bind(&b.size)
        .fetch_optional(&mut *tx)
        .await?;
        tx.commit().await?;
        let out = row.map(|r| r.0).unwrap_or_else(|| {
            json!({
                "brand": b.brand,
                "model": b.model.clone().unwrap_or_default(),
                "size": b.size.clone().unwrap_or_default(),
                "on_hand_qty": 0,
            })
        });
        return Ok(HttpResponse::Ok().json(out));
    }
    let row: (Value,) = sqlx::query_as(
        r#"
        INSERT INTO tire_stock_cache (brand, model, size, on_hand_qty)
        VALUES ($1, COALESCE($2,''), COALESCE($3,''), $4)
        ON CONFLICT (brand, model, size) DO UPDATE SET
            on_hand_qty = tire_stock_cache.on_hand_qty + EXCLUDED.on_hand_qty,
            fetched_at = now()
        RETURNING to_jsonb(tire_stock_cache)
        "#,
    )
    .bind(&b.brand)
    .bind(&b.model)
    .bind(&b.size)
    .bind(b.qty)
    .fetch_one(&mut *tx)
    .await?;
    tx.commit().await?;
    Ok(HttpResponse::Ok().json(row.0))
}

fn parse_opt_uuid(s: Option<&str>) -> ApiResult<Option<uuid::Uuid>> {
    match s {
        None => Ok(None),
        Some(s) => uuid::Uuid::parse_str(s)
            .map(Some)
            .map_err(|_| ApiError::BadRequest("ref_id must be a uuid".into())),
    }
}

#[derive(Debug, Deserialize)]
struct OilCredit {
    oil_type: String,
    liters: f64,
    /// Idempotency key — see TireCredit::ref_id.
    ref_id: Option<String>,
}

#[post("/stock/oil/credit")]
async fn credit_oil(
    pool: web::Data<PgPool>,
    claims: AuthClaims,
    body: web::Json<OilCredit>,
) -> ApiResult<HttpResponse> {
    if claims.permission < 4 {
        return Err(ApiError::Forbidden("stock credit requires permission >= 4".into()));
    }
    let b = body.into_inner();
    if b.liters <= 0.0 || b.liters.is_nan() {
        return Err(ApiError::BadRequest("liters must be > 0".into()));
    }
    let ref_id = parse_opt_uuid(b.ref_id.as_deref())?;

    let mut tx = pool.begin().await?;
    let logged: Option<(uuid::Uuid,)> = sqlx::query_as(
        "INSERT INTO stock_movements (kind, oil_type, liters, ref_id) \
         VALUES ('oil_credit', $1, $2, COALESCE($3, gen_random_uuid())) \
         ON CONFLICT (ref_id) WHERE ref_id IS NOT NULL DO NOTHING RETURNING id",
    )
    .bind(&b.oil_type)
    .bind(b.liters)
    .bind(ref_id)
    .fetch_optional(&mut *tx)
    .await?;
    if logged.is_none() {
        // See credit_tires: idempotency is keyed on ref_id, so a replay may name
        // a different oil_type than the original. fetch_OPTIONAL + zero fallback.
        let row: Option<(Value,)> =
            sqlx::query_as("SELECT to_jsonb(t) FROM oil_stock_cache t WHERE oil_type=$1")
                .bind(&b.oil_type)
                .fetch_optional(&mut *tx)
                .await?;
        tx.commit().await?;
        let out = row.map(|r| r.0).unwrap_or_else(|| {
            json!({ "oil_type": b.oil_type, "liters_on_hand": 0.0 })
        });
        return Ok(HttpResponse::Ok().json(out));
    }
    let row: (Value,) = sqlx::query_as(
        r#"
        INSERT INTO oil_stock_cache (oil_type, liters_on_hand)
        VALUES ($1, $2)
        ON CONFLICT (oil_type) DO UPDATE SET
            liters_on_hand = oil_stock_cache.liters_on_hand + EXCLUDED.liters_on_hand,
            fetched_at = now()
        RETURNING to_jsonb(oil_stock_cache)
        "#,
    )
    .bind(&b.oil_type)
    .bind(b.liters)
    .fetch_one(&mut *tx)
    .await?;
    tx.commit().await?;
    Ok(HttpResponse::Ok().json(row.0))
}

/// The in/out ledger: every recorded stock movement (debits from oil changes
/// and shipment-tire mounts, credits from the dashboard), newest first, with
/// the consuming work order's plate resolved. `kind` filters to tire | oil.
#[derive(Debug, Deserialize)]
struct MovementsQuery {
    kind: Option<String>,
    /// Inclusive lower bound on created_at (date or timestamp) — the
    /// time-range filter. Absent = no lower bound.
    from: Option<String>,
    limit: Option<i64>,
    offset: Option<i64>,
}

#[get("/stock/movements")]
async fn stock_movements(
    pool: web::Data<PgPool>,
    _claims: AuthClaims,
    q: web::Query<MovementsQuery>,
) -> ApiResult<HttpResponse> {
    let limit = q.limit.unwrap_or(50).clamp(1, 200);
    let offset = q.offset.unwrap_or(0).max(0);
    let kind_filter = match q.kind.as_deref() {
        Some("tire") => " AND m.kind IN ('tire_debit','tire_credit')",
        Some("oil") => " AND m.kind IN ('oil_debit','oil_credit')",
        _ => "",
    };
    // `from` binds as $3 (Option → NULL when absent). A bad timestamp casts to
    // a SQLSTATE-22 error, which error.rs maps to a clean 400, not a 500.
    let sql = format!(
        "SELECT jsonb_build_object(
             'id', m.id, 'kind', m.kind::text, 'brand', m.brand, 'model', m.model, 'size', m.size,
             'oil_type', m.oil_type, 'qty', m.qty, 'liters', m.liters,
             'work_order_id', m.work_order_id, 'plate', v.car_no_plate,
             -- Render in explicit UTC so the string is constant-offset and the
             -- client's offline lexical ORDER BY created_at is always correct
             -- (a Cairo-DST session tz would otherwise vary the offset).
             'created_at', to_char(m.created_at AT TIME ZONE 'UTC', 'YYYY-MM-DD\"T\"HH24:MI:SS.US\"Z\"'),
             'mirrored', (m.mirrored_to_falcon_at IS NOT NULL))
           FROM stock_movements m
           LEFT JOIN work_orders wo ON wo.id = m.work_order_id
           LEFT JOIN vehicles_cache v ON v.id = wo.vehicle_id
          WHERE ($3::timestamptz IS NULL OR m.created_at >= $3::timestamptz){kind_filter}
          ORDER BY m.created_at DESC, m.id DESC
          LIMIT $1 OFFSET $2"
    );
    let rows: Vec<(Value,)> = sqlx::query_as(&sql)
        .bind(limit)
        .bind(offset)
        .bind(&q.from)
        .fetch_all(pool.get_ref())
        .await?;
    Ok(HttpResponse::Ok().json(rows.into_iter().map(|(v,)| v).collect::<Vec<_>>()))
}

/// Dev-only: seed a batch of tire + oil stock so the garage has something to consume.
#[post("/stock/dev-seed")]
async fn dev_seed(state: web::Data<AppState>, pool: web::Data<PgPool>) -> ApiResult<HttpResponse> {
    if !cfg!(debug_assertions) || !state.dev_login {
        return Err(ApiError::Forbidden("dev-seed disabled".into()));
    }
    sqlx::query(
        "INSERT INTO tire_stock_cache (brand, model, size, on_hand_qty) VALUES \
         ('Pirelli', 'FR01', '315/80R22.5', 20), ('Tegrys', 'T20', '315/80R22.5', 10) \
         ON CONFLICT (brand, model, size) DO UPDATE SET on_hand_qty = EXCLUDED.on_hand_qty",
    )
    .execute(pool.get_ref())
    .await?;
    sqlx::query(
        "INSERT INTO oil_stock_cache (oil_type, liters_on_hand) VALUES \
         ('15W-40', 400), ('80W-90', 120) \
         ON CONFLICT (oil_type) DO UPDATE SET liters_on_hand = EXCLUDED.liters_on_hand",
    )
    .execute(pool.get_ref())
    .await?;
    // A few dev vehicles so the app has trucks to service without Falcon.
    sqlx::query(
        "INSERT INTO vehicles_cache (id, car_no_plate, car_type, transporter, last_fuel_odometer, raw_payload) VALUES \
         (101, 'ن ص ر 1234', 'تريلا', 'Apex', 250000, '{}'::jsonb), \
         (102, 'ب ح ر 5678', 'تريلا', 'Apex', 180500, '{}'::jsonb), \
         (103, 'ق ا د 9012', 'No Trailer', 'Apex', 320000, '{}'::jsonb) \
         ON CONFLICT (id) DO UPDATE SET car_no_plate = EXCLUDED.car_no_plate",
    )
    .execute(pool.get_ref())
    .await?;
    Ok(HttpResponse::Ok().json(serde_json::json!({"seeded": true})))
}

pub fn configure(cfg: &mut web::ServiceConfig) {
    cfg.service(tire_stock)
        .service(oil_stock)
        .service(stock_movements)
        .service(debit_stock)
        .service(credit_tires)
        .service(credit_oil)
        .service(dev_seed);
}